feat: began new idea for blog and created template for future blogs
A thought came to me while reading "Where We Want to Live" by Ryan Gravel. If I want to see EV Charging here in GA, I have to present an idea and spread it like wildfire. Nothing will happen unless there is excitement about the prospect of something potentially wonderful. The Beltline was just an idea of a Master's student who just wanted to get out the door with a good thesis. They didn't think it would materialize. Well, let's build an idea with intent and see who might buy into it. Make it appealing and share it with city leaders so they have a vision of what tomorrow could look like.